Midterm/services/requirements.txt
requests>=2.32.3
python-dotenv>=1.1.0
youtube-transcript-api>=1.0.3
pydantic>=2.11.4
mongoengine>=0.29.1
fastapi[standard]>=0.115.12
openai>=1.78.1
pytest>=8.3.5
pytest-cov>=6.1.1
mypy>=1.15.0
ruff>=0.11.9
httpx>=0.28.1
supabase>=2.3.5 